Fix LOAD segment with RWX permissions error
Newer toolchain complains about loaded segments with RWX permissions.

Our linker script was actually correct.  But we had a couple of
erroneous const void * (pointer to constant void) instead of void *
const (const pointer to void) in the interrupt vector table.  As a
result, the vectors were marked writable.  They were adjacent to code
and text, so the whole thing was rwx as a result.

Now, it's properly r-x.