improve branch calculator and fix some bugs
While this is still as-yet untested, I noticed that the subtraction
was in the incorrect order (was start - target instead of the
opposite) and that I had failed to account for the way the
instruction pointer is already incremented at the time of the
branch.

While we're at it, let's make sure the source address is valid,
and construct a branch instruction ready for the user to hit
store.